+++ src/Type/Collection/Interfaces/Type/NumericCollectionInterface.php
@@ -3,8 +3,9 @@
 namespace LDL\Type\Collection\Interfaces\Type;
 
 use LDL\Type\Collection\Interfaces\Type\ToPrimitiveArrayInterface;
+use LDL\Type\Collection\Interfaces\Type\NumericTypeCollectionInterface;
 
-interface NumericCollectionInterface extends ToPrimitiveArrayInterface
+interface NumericCollectionInterface extends ToPrimitiveArrayInterface, NumericTypeCollectionInterface
 {
     
 }