+++ src/Type/Collection/Types/Integer/UnsignedIntegerCollection.php
@@ -7,11 +7,13 @@ use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Interfaces\Type\IntegerCollectionInterface;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
 use LDL\Type\Collection\Types\Integer\Traits\ToIntegerPrimitiveArrayTrait;
+use LDL\Type\Collection\Types\Integer\Traits\IntegerCollectionInterfaceTrait;
 use LDL\Type\Collection\Interfaces\Validation\HasAppendValueValidatorChainInterface;
 use LDL\Type\Collection\Traits\Types\Integer\FilterUniqueIntegerCollectionInterfaceTrait;
 
 final class UnsignedIntegerCollection extends AbstractTypedCollection implements IntegerCollectionInterface
 {
+    use IntegerCollectionInterfaceTrait;
     use FilterUniqueIntegerCollectionInterfaceTrait;
     use ToIntegerPrimitiveArrayTrait;